+++ src/lib/providers/IProvider.ts
@@ -1,0 +1,7 @@
+import { BigNumber } from 'ethers';
+
+export interface IProvider {
+  getBlockNumber(): Promise<number>;
+  getBalance(address: string): Promise<BigNumber>;
+  getTransactionCount(address: string): Promise<number>;
+}