+++ src/repositories/ChainContractRepository.ts
@@ -1,12 +1,14 @@
 import { inject, injectable } from 'inversify';
 import Settings from '../types/Settings';
-import { ChainsIds, ForeignChainsIds } from '../types/ChainsIds';
+import { ChainsIds, ForeignChainsIds, NonEvmChainsIds } from '../types/ChainsIds';
 import { ChainContract } from '../contracts/ChainContract';
 import { BlockchainRepository } from './BlockchainRepository';
 import { ForeignChainContract } from '../contracts/ForeignChainContract';
+import { IGenericForeignChainContract } from '../contracts/generic/IGenericForeignChainContract';
+import { ForeignChainContractFactory } from '../factories/ForeignChainContractFactory';
 
 export type ChainContractCollection = {
-  [key: string]: ChainContract | ForeignChainContract;
+  [key: string]: ChainContract | ForeignChainContract | IGenericForeignChainContract;
 };
 
 @injectable()
@@ -25,12 +27,17 @@ export class ChainContractRepository {
         : undefined;
 
     ForeignChainsIds.forEach((foreignChainId) => {
-      const blockchain = blockchainRepository.get(foreignChainId);
+      if (NonEvmChainsIds.includes(foreignChainId)) {
+        const blockchain = blockchainRepository.getGeneric(foreignChainId);
+        this.collection[foreignChainId] = ForeignChainContractFactory.create({ blockchain, settings });
+      } else {
+        const blockchain = blockchainRepository.get(foreignChainId);
 
-      this.collection[foreignChainId] =
-        blockchain.provider && blockchain.getContractRegistryAddress()
-          ? new ForeignChainContract({ blockchain, settings })
-          : undefined;
+        this.collection[foreignChainId] =
+          blockchain.provider && blockchain.getContractRegistryAddress()
+            ? new ForeignChainContract({ blockchain, settings })
+            : undefined;
+      }
     });
   }
 
@@ -39,6 +46,22 @@ export class ChainContractRepository {
       console.warn(`[ChainContractRepository] Blockchain ${id} does not exists`);
     }
 
-    return this.collection[id];
+    if (NonEvmChainsIds.includes(<ChainsIds>id)) {
+      throw Error(`[ChainContractRepository] Wrong Blockchain type for ${id}`);
+    }
+
+    return <ChainContract | ForeignChainContract>this.collection[id];
+  }
+
+  getGeneric(id: string): IGenericForeignChainContract {
+    if (!this.collection[id]) {
+      throw Error(`[ChainContractRepository] Blockchain ${id} does not exists`);
+    }
+
+    if (!NonEvmChainsIds.includes(<ChainsIds>id)) {
+      throw Error(`[ChainContractRepository] Wrong Blockchain type for ${id}. Expected GenericBlockchain`);
+    }
+
+    return <IGenericForeignChainContract>this.collection[id];
   }
 }